pyproject.toml
[tool.poetry]
name = "projeto eleições"
version = "0.1.0"
description = ""
authors = ["GabMeula07 <gabrielqueirozmeula@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
env = "^0.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
chardet = "^5.2.0"
openpyxl = "^3.1.5"
sqlalchemy = "^2.0.35"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"
ipykernel = "^6.29.5"

[tool.ruff.format]
indent-style = "space"

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = ["E", "F"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
